주석 수정, 코드 가독성 수정, 상점삭제시 자신의 상점만 삭제하도록 코드 수정

+++ src/main/java/com/example/tablereservation/store/service/StoreService.java
@@ -69,10 +69,17 @@ public class StoreService {
 
     /**
      * 상점 삭제
+     * 1. 상점이 존재하는지 확인
+     * 2. 본인의 상점인지 확인
      */
-    public void deleteStore(Long storeId) {
+    public void deleteStore(Long storeId, PartnerEntity partnerEntity) {
         StoreEntity storeEntity = this.storeRepository.findById(storeId)
                 .orElseThrow(() -> new ReservationException(ErrorCode.STORE_NOT_EXIST));
+
+        if (!storeEntity.getPartner().getId().equals(partnerEntity.getId())) {
+            throw new ReservationException(ErrorCode.UNAUTHORIZED);
+        }
+
         storeRepository.delete(storeEntity);
     }
